restructures and adds styles for Tile
