--- src/__tests__/Badge.test.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Badge from '../components/Badge';
-
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<Badge label="This is a badge" />, div);
-});
