--- src/__tests__/Footer.test.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import {render} from 'react-dom';
-import Footer from '../components/Footer';
-
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  render(<Footer />, div);
-});
