--- src/components/Badge.scss
@@ -1,1 +1,0 @@
-@import "../styles/components/badge";
