+++ src/components/Badge/__tests__/Badge.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Badge from '../../Badge';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<Badge label="This is a badge" />, div);
+});
