--- src/components/Button.scss
@@ -1,1 +1,0 @@
-@import "../styles/components/buttons";
