+++ src/components/Button/__tests__/Button.test.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { render } from "react-dom";
+import Button from "../../Button";
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  render(<Button type="submit" handleClick={function(){}} />, div);
+});
