--- src/components/Footer.scss
@@ -1,1 +1,0 @@
-@import "../styles/components/footer";
