+++ src/components/Footer/__tests__/Footer.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import {render} from 'react-dom';
+import Footer from '../../Footer';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  render(<Footer />, div);
+});
