--- src/components/Header.scss
@@ -1,1 +1,0 @@
-@import "../styles/components/header";
