+++ src/components/Nav/styles.scss
@@ -1,0 +1,1 @@
+@import "../../styles/components/nav";
