--- src/components/SkipLink.scss
@@ -1,1 +1,0 @@
-@import "../styles/components/skip-link";
