+++ src/components/SkipLink/styles.scss
@@ -1,0 +1,1 @@
+@import "../../styles/components/skip-link";
