--- src/components/SocialIcons.scss
@@ -1,1 +1,0 @@
-@import "../styles/components/social-icons"
