+++ src/components/SocialIcons/__tests__/SocialIcons.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import {render} from 'react-dom';
+import SocialIcons from '../../SocialIcons';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  render(<SocialIcons />, div);
+});
