+++ src/components/SocialIcons/styles.scss
@@ -1,0 +1,1 @@
+@import "../../styles/components/social-icons"
