+++ src/styles/components/_nav.scss
@@ -1,0 +1,30 @@
+@import "../styles";
+
+.nav {
+  text-align: right;
+  text-transform: uppercase;
+  @include grid-column(5);
+  @media screen and (min-width: $break-md) {
+    @include grid-column(6);
+    float: right;
+  }
+  @media screen and (min-width: $break-lg) {
+    @include grid-column(5);
+  }
+}
+
+.nav__list {margin: 0;}
+.nav__item {
+  display: inline-block;
+  &:hover {
+    background-color: $deep-navy-lighten;
+    transition: $transition-default
+  }
+  @media screen and (min-width: $break-md) {
+    padding-left: 14px;
+    padding-right: 14px;
+  }
+}
+.nav__link {
+  @include secondary-header-nav-link;
+}
